Resume-Relevance-Scoring-master/backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cryptography==42.0.2
cymem==2.0.8
distro==1.9.0
python-dotenv==1.0.1
Django==3.2.24
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
Jinja2==3.1.3
langcodes==3.3.0
MarkupSafe==2.1.5
murmurhash==1.0.10
numpy
openai==1.12.0
packaging==23.2
pdfminer.six==20231228
pillow==10.2.0
preshed==3.0.9
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
pytesseract==0.3.10
pytz==2024.1
requests==2.31.0
smart-open==6.4.0
sniffio==1.3.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
thinc==8.2.3
tqdm==4.66.2
typer==0.9.0
typing_extensions==4.9.0
tiktoken==0.6.0
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4